Added wwise project to our engine, so everyone can add sound events to the soundbank.
All that you need to do is open up the project. Import a file. Make an event off of that sound. Open up the soundbank window. Double click the titanswithsticks bank. Drag the event over to the add section. Generate the soundbank.
Now that the soundbank is generated, go to the place you want to play a sound and look at my sound code for reference.